easy coding -
recap & re-organized about 排序算法.

1 插入排序 - 把第一个元素作为有序区，逐步扩展有序区到整个数组；
2 归并排序 - 以有序数组 与 归并操作为基础，逐步得到一个个完全有序的局部区间，最终得到整个有序数组；
3 TimSort - do better.
