+++ src/main/java/com/henry/tryout/easy_coding/data_structure_and_collection/compare_items_06/sort_algorithm_02/sort_algorithms
@@ -3,15 +3,15 @@ Arrays.sort() 使用的排序方式 - TimSort
 TimSort = Merge Sort + Insertion Sort；
 
 归并排序；
-    归并操作：对于两个已经排好序的数组a与数组b，通过归并操作，能够得到一个完全排序的数组c。
+    归并操作：对于两个 已经排好序的 数组a与数组b，通过"归并操作"，就能够得到一个 完全排序的数组c。
     步骤：
         #1 n个元素的集合 等价于 n个长度为1的有序数组；
-        #2 对有序的子数组进行两两归并，这样会得到 n/2个 长度为2的有序数组；
+        #2 对 有序的子数组 进行两两归并，这样会得到 n/2个 长度为2的有序数组；
         #3 重复步骤2，直到得到一个 长度为n的有序数组。
-    特征：主要的耗时都在归并操作上。
+    特征：主要的耗时都在 归并操作 上。
 
 插入排序；
-    #1 维护一个有序的区域，然后把下一个待处理的数据插入到该有序区域中。
+    #1 维护一个有序的区域，然后把 下一个待处理的数据 插入到该有序区域中。
     #2 扩展该有序区域，直到所有的元素都已经有序。
 
 Tim Peters/Sort;
@@ -29,17 +29,3 @@ Tim Peters/Sort;
         二分排序是对插入排序的优化：
             不再是从后向前逐个地对比元素，而是使用二分查找的思想。
             这样查找合适的插入位置的时间复杂度 从O(n) -> O(logn);
-
-======
-hashCode 与 equals
-
-作用：标识对象；
-    这两个方法协同工作来判断两个对新啊给是否相等；
-原理：
-    1 调用Object.hashCode()生成hash值时，很有可能会产生哈希冲突；
-    2 所以在 hashCode相同的情况下，还需要再调用equals() 进行一次值的比较。
-规则：
-    #1 如果两个对象的 equals()是相等的，那么 两个对象调用hashCode()返回的结果也必须是相同的；
-    #2 任何覆写 equals 的时候，都需要同时覆写 hashCode.
-应用：
-    1 HashMap get()方法；
